requirements.txt
mediapipe
numpy
pandas
opencv-python
cvzone
google-generativeai
python-dotenv
streamlit
flask
